requirements.txt
faker==40.1.0
psycopg2-binary==2.9.9
pandas==2.2.0
openpyxl==3.1.2
